Fixed a crash occurring when holding an item and clicking between the slots.
I also realized I had originally changed the GridContainer's mouse filter to Pass because I read the README, and forgot to change it before. Not everyone is going to read the README, so I thought it best to change that.

I also stuck a semicolon at the end of the return lines to keep with the original style.